+++ CHANGELOG.md
@@ -23,8 +23,8 @@
 - Automatically Escape `'` in the _text_ editors in the `In-Editor Parameters` view [#1454](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1454)
 
 ### `Polyglot Query` enhancements
-
 - Introduced `In-Editor Results` view [#1462](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1462)
+- Introduced `In-Editor Parameters` view [#1463](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1463)
 
 ### `Groovy` enhancements
 - Introduced `batch` execution mode for Groovy files [#1450](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1450)